test: migrate zero-maps to bdd
